+++ src/pubsub/pub_sub_config.ts
@@ -14,5 +14,11 @@ export const PubSubConfig = {
     format: String,
     default: process.env.NODE_ENV,
     env: 'GCP_PUBSUB_TOPIC_PREFIX'
+  },
+  resetHostedSubscriptions: {
+    doc: 'Should all hosted subscriptions be reset on startup?',
+    format: Boolean,
+    default: false,
+    env: 'GCP_PUBSUB_RESET_HOSTED_SUBSCRIPTIONS'
   }
 };